pyproject.toml
[project]
name = "simple-aigc-detect"
version = "0.1.0"
description = "A CLI tool to detect AIGC generated content in documents"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "transformers",
    "torch",
    "python-docx",
    "typer",
    "rich",
    "huggingface-hub"
]

[project.scripts]
aigc-detect = "simple_aigc_detect.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["simple_aigc_detect"]
